% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{as_pkgsite}
\alias{as_pkgsite}
\title{Generate minimal pkgsite data structure}
\usage{
as_pkgsite(pkg = ".")
}
\arguments{
\item{pkg}{Path to package.}
}
\description{
Generate minimal pkgsite data structure
}
