% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-home.R
\name{build_home}
\alias{build_home}
\alias{build_home_index}
\title{Build home section}
\usage{
build_home(pkg = ".", quiet = FALSE, preview = FALSE)

build_home_index(pkg = ".", quiet = FALSE)
}
\arguments{
\item{pkg}{Path to package}

\item{quiet}{Whether to suppress messages}

\item{preview}{Whether to preview after building}
}
\description{
\code{build_home()} generates the home page from README.md or index.md
}
