% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-news.R
\name{build_news}
\alias{build_news}
\title{Build news page}
\usage{
build_news(pkg = ".", quiet = TRUE, preview = FALSE)
}
\arguments{
\item{pkg}{Path to package to document.}

\item{quiet}{If \code{quiet}, will suppress output messages}

\item{preview}{If \code{TRUE}, the news page will be opened in browser}
}
\description{
Generates a news page from NEWS.md, NEWS.Rmd, or ChangeLog files.
}
