% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-reference.R
\name{build_reference}
\alias{build_reference}
\alias{build_reference_index}
\title{Build reference section}
\usage{
build_reference(pkg = ".", lazy = FALSE, preview = FALSE)

build_reference_index(pkg = ".")
}
\arguments{
\item{pkg}{Path to package}

\item{lazy}{Only rebuild if source is newer than destination}

\item{preview}{Whether to preview after building}
}
\description{
\code{build_reference()} generates reference documentation from Rd files
}
