% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-search.R
\name{build_search_index}
\alias{build_search_index}
\title{Build search index}
\usage{
build_search_index(pkg = ".")
}
\arguments{
\item{pkg}{Path to package}
}
\description{
Creates a search index of all content for client-side searching
}
