% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown-utils.R
\name{clean_rd_content}
\alias{clean_rd_content}
\title{Clean Rd content by removing common LaTeX commands}
\usage{
clean_rd_content(content)
}
\arguments{
\item{content}{Character string containing Rd content}
}
\value{
Character string with LaTeX commands cleaned
}
\description{
Clean Rd content by removing common LaTeX commands
}
