% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown-utils.R
\name{clean_text_for_search}
\alias{clean_text_for_search}
\title{Clean search text by removing markdown and HTML}
\usage{
clean_text_for_search(text)
}
\arguments{
\item{text}{Character string containing formatted text}
}
\value{
Character string with formatting removed for search indexing
}
\description{
Clean search text by removing markdown and HTML
}
