% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown-utils.R
\name{convert_angle_bracket_urls}
\alias{convert_angle_bracket_urls}
\title{Convert angle-bracket URLs to HTML links}
\usage{
convert_angle_bracket_urls(content)
}
\arguments{
\item{content}{Character string containing markdown content}
}
\value{
Character string with angle-bracket URLs converted to HTML
}
\description{
Convert angle-bracket URLs to HTML links
}
