% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown-utils.R
\name{convert_code_blocks_simple}
\alias{convert_code_blocks_simple}
\title{Convert markdown code blocks to HTML (simple regex method)}
\usage{
convert_code_blocks_simple(content, language_specific = FALSE)
}
\arguments{
\item{content}{Character string containing markdown content}

\item{language_specific}{Whether to preserve language specifications}
}
\value{
Character string with code blocks converted to HTML
}
\description{
A simpler approach using regex for cases where order doesn't matter
}
