% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown-utils.R
\name{convert_inline_code}
\alias{convert_inline_code}
\title{Convert markdown inline code to HTML}
\usage{
convert_inline_code(content)
}
\arguments{
\item{content}{Character string containing markdown content}
}
\value{
Character string with inline code converted to HTML
}
\description{
Convert markdown inline code to HTML
}
