% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown-utils.R
\name{convert_markdown_headers}
\alias{convert_markdown_headers}
\title{Convert markdown headers to HTML}
\usage{
convert_markdown_headers(content)
}
\arguments{
\item{content}{Character string containing markdown content}
}
\value{
Character string with headers converted to HTML
}
\description{
Convert markdown headers to HTML
}
