% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown-utils.R
\name{convert_markdown_headers_simple}
\alias{convert_markdown_headers_simple}
\title{Convert markdown headers to HTML (simple regex method)}
\usage{
convert_markdown_headers_simple(content)
}
\arguments{
\item{content}{Character string containing markdown content}
}
\value{
Character string with headers converted to HTML
}
\description{
A simpler approach using regex for multiline content
}
