% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown-utils.R
\name{convert_markdown_images}
\alias{convert_markdown_images}
\title{Convert markdown images to HTML}
\usage{
convert_markdown_images(content)
}
\arguments{
\item{content}{Character string containing markdown content}
}
\value{
Character string with images converted to HTML
}
\description{
Convert markdown images to HTML
}
