% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown-utils.R
\name{convert_markdown_links}
\alias{convert_markdown_links}
\title{Convert markdown links to HTML}
\usage{
convert_markdown_links(content)
}
\arguments{
\item{content}{Character string containing markdown content}
}
\value{
Character string with links converted to HTML
}
\description{
Convert markdown links to HTML
}
