% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown-utils.R
\name{convert_markdown_lists}
\alias{convert_markdown_lists}
\title{Convert markdown lists to HTML}
\usage{
convert_markdown_lists(content)
}
\arguments{
\item{content}{Character string containing markdown content}
}
\value{
Character string with lists converted to HTML
}
\description{
Convert markdown lists to HTML
}
