% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown-utils.R
\name{convert_rd_to_html}
\alias{convert_rd_to_html}
\title{Convert common Rd commands to HTML}
\usage{
convert_rd_to_html(text)
}
\arguments{
\item{text}{Character string containing Rd text}
}
\value{
Character string with Rd commands converted to HTML
}
\description{
Convert common Rd commands to HTML
}
