% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{data_template}
\alias{data_template}
\title{Generate template data}
\usage{
data_template(pkg = ".", root_path = "")
}
\arguments{
\item{pkg}{Path to package}

\item{root_path}{Root path for relative links}
}
\value{
List with template data
}
\description{
Internal function to generate template data for rendering pages
}
