% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown-utils.R
\name{extract_rd_section}
\alias{extract_rd_section}
\title{Extract Rd section content}
\usage{
extract_rd_section(content, section_name)
}
\arguments{
\item{content}{Character string containing Rd content}

\item{section_name}{Name of the section to extract (e.g., "title", "description")}
}
\value{
Character string with section content
}
\description{
Extract Rd section content
}
