% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_site}
\alias{init_site}
\title{Initialise minimal site infrastructure}
\usage{
init_site(pkg = ".")
}
\arguments{
\item{pkg}{Path to package}
}
\description{
\code{init_site()} creates the output directory and copies basic CSS files.
}
