% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown-utils.R
\name{markdown_to_html_full}
\alias{markdown_to_html_full}
\title{Full markdown to HTML conversion}
\usage{
markdown_to_html_full(content, language_specific = FALSE)
}
\arguments{
\item{content}{Character string containing markdown content}

\item{language_specific}{Whether to preserve language specifications in code blocks}
}
\value{
Character string with markdown converted to HTML
}
\description{
Applies all markdown conversions in the correct order
}
