% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-to-list.R
\name{rd_to_list}
\alias{rd_to_list}
\title{Converts a given 'Rd' file into a list object}
\usage{
rd_to_list(path)
}
\arguments{
\item{path}{The path to the source Rd file}
}
\value{
A list object that contains the contents of the Rd file
}
\description{
This function is meant to be used as an intermediate object
that could be used as an easy way to convert the information inside the 'Rd'
into other formats or outputs.
}
\examples{
library(pkgsite)
example_pkg <- system.file("example", package = "pkgsite")
rd_to_list(file.path(example_pkg, "man", "rd_to_list.Rd"))
}
\seealso{
Other Conversion functions:
\code{\link[=index_to_qmd]{index_to_qmd()}},
\code{\link[=rd_to_qmd]{rd_to_qmd()}}
}
\concept{Conversion functions}
