% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-to-qmd.R
\name{rd_to_qmd}
\alias{rd_to_qmd}
\title{Converts 'Rd' to Quarto files}
\usage{
rd_to_qmd(
  path,
  pkg = ".",
  examples = TRUE,
  not_run_examples = FALSE,
  template = NULL,
  quarto_file_path = NULL
)
}
\arguments{
\item{path}{The path to the source Rd file}

\item{pkg}{Path to the root of the R package. Defaults to \code{"."}.}

\item{examples}{Flag that sets the examples code chunk to be evaluated when
the Quarto document is rendered}

\item{not_run_examples}{Flag that sets the "do not run" examples code chunk
to be evaluated when the Quarto document is rendered}

\item{template}{The path to a Quarto file that can be used as the template
for all of the resulting reference files. If left NULL, \code{pkgsite} will use
its default template.}

\item{quarto_file_path}{Path to the folder containing \verb{_quarto.yml}. Use
when the Quarto project root differs from the package root (e.g. the
package lives in a sub-folder of a larger Quarto project). Must be a
directory path. Defaults to \code{NULL}, in which case \code{pkg} is used.}
}
\value{
A character vector with the resulting contents of converting the
Rd file format into a Quarto file format.
}
\description{
Converts 'Rd' to Quarto files
}
\examples{
library(pkgsite)
example_pkg <- system.file("example", package = "pkgsite")
rd_to_qmd(file.path(example_pkg, "man", "rd_to_qmd.Rd"), pkg = example_pkg)
}
\seealso{
Other Conversion functions:
\code{\link[=index_to_qmd]{index_to_qmd()}},
\code{\link[=rd_to_list]{rd_to_list()}}
}
\concept{Conversion functions}
