% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_page}
\alias{render_page}
\alias{get_existing_reference_topics}
\title{Render page with minimal template}
\usage{
render_page(pkg = ".", name, data, path, quiet = FALSE)

get_existing_reference_topics(pkg)
}
\arguments{
\item{pkg}{Path to package to document.}

\item{name}{Name of the template (e.g. "home", "reference-topic")}

\item{data}{Data for the template.}

\item{path}{Location to create file; relative to destination directory.}

\item{quiet}{If \code{quiet}, will suppress output messages}

\item{root_path}{Relative path to site root (automatically calculated)}
}
\description{
Each page uses simple templates: "head", "header", "content", and "footer".
}
