% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown-utils.R
\name{wrap_paragraphs}
\alias{wrap_paragraphs}
\title{Wrap content in paragraphs}
\usage{
wrap_paragraphs(content)
}
\arguments{
\item{content}{Character string containing HTML content}
}
\value{
Character string with paragraphs wrapped
}
\description{
Takes content and wraps text that isn't already HTML elements in \if{html}{\out{<p>}} tags
}
