% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-site.R
\name{write_last_updated}
\alias{write_last_updated}
\title{Write last updated timestamp}
\usage{
write_last_updated(pkg = ".")
}
\arguments{
\item{pkg}{Path to package}
}
\description{
Writes a JSON file with the last updated timestamp to avoid
embedding it in every HTML file, which causes large git diffs
}
\keyword{internal}
