% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-reference.R
\name{write_reference_pages}
\alias{write_reference_pages}
\title{Converts the 'Rd' file into Quarto, and writes the file to a specified folder}
\usage{
write_reference_pages(
  pkg = ".",
  folder = "reference",
  examples = TRUE,
  not_run_examples = FALSE,
  template = NULL
)
}
\arguments{
\item{pkg}{The path to the package or a \code{pkgdown} package object}

\item{folder}{The target folder to save the new Quarto files to}

\item{examples}{Flag that sets the examples code chuck to be evaluated when
the Quarto document is rendered}

\item{not_run_examples}{Flag that sets the "do not run" examples code chuck
to be evaluated when the Quarto document is rendered}

\item{template}{The path to a Quarto file that can be used as the template
for all of the resulting reference files. If left NULL, \code{pkgsite} will use
its default template.}
}
\description{
Converts the 'Rd' file into Quarto, and writes the file to a specified folder
}
