% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-news.R
\name{build_news}
\alias{build_news}
\title{Build news page}
\usage{
build_news(pkg = ".", quiet = TRUE, preview = FALSE)
}
\arguments{
\item{pkg}{Path to package to document.}

\item{quiet}{If \code{quiet}, will suppress output messages}

\item{preview}{If \code{TRUE}, the news page will be opened in browser}
}
\value{
Invisible \code{TRUE} if the news page was built successfully
}
\description{
Generates a news page from NEWS.md
}
\examples{
\dontrun{
build_news()

# Hide build messages
build_news(quiet = TRUE)

# Preview the news page after building
build_news(preview = TRUE)
}

}
