% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-reference.R
\name{build_reference}
\alias{build_reference}
\alias{build_reference_index}
\title{Build reference section}
\usage{
build_reference(pkg = ".", lazy = FALSE, preview = FALSE)

build_reference_index(pkg = ".")
}
\arguments{
\item{pkg}{Path to package}

\item{lazy}{Only rebuild if source is newer than destination}

\item{preview}{Whether to preview after building}
}
\value{
Invisible \code{TRUE} if the reference was built successfully
}
\description{
\code{build_reference()} generates reference documentation from Roxygen comments
}
\examples{
\dontrun{
build_reference()

# Only rebuild if source is newer than destination
build_reference(lazy = TRUE)

# Preview the site after building
build_reference(preview = TRUE)
}

}
