% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-vignettes.R
\name{build_vignettes}
\alias{build_vignettes}
\alias{build_vignettes_index}
\title{Build vignettes section}
\usage{
build_vignettes(pkg = ".", lazy = FALSE, quiet = FALSE, preview = FALSE)

build_vignettes_index(pkg = ".")
}
\arguments{
\item{pkg}{Path to package}

\item{lazy}{Whether to skip files that haven't changed}

\item{quiet}{Whether to suppress messages}

\item{preview}{Whether to preview after building using a local server (requires the \code{servr} package)}
}
\value{
Invisible \code{TRUE} if the vignettes were built successfully
}
\description{
\code{build_vignettes()} renders R Markdown files from vignettes/ to vignettes/
}
\examples{
\dontrun{
build_vignettes()

# Rebuild only if source is newer than destination
build_vignettes(lazy = TRUE)

# Hide build messages
build_vignettes(quiet = TRUE)

# Preview the vignettes index after building
build_vignettes(preview = TRUE)
}

}
