% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown-utils.R
\name{convert_code_blocks}
\alias{convert_code_blocks}
\title{Convert markdown code blocks to HTML}
\usage{
convert_code_blocks(content, language_specific = FALSE)
}
\arguments{
\item{content}{Character string containing markdown content}

\item{language_specific}{Whether to preserve language specifications}
}
\description{
Convert markdown code blocks to HTML
}
