% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_to_qmd.R
\name{index_to_qmd}
\alias{index_to_qmd}
\title{Create a Quarto file that lists the available reference pages}
\usage{
index_to_qmd(project = ".", pkg = NULL, template = NULL)
}
\arguments{
\item{project}{The path to the root folder of the project.}

\item{pkg}{The path inside the project folder. Use only if the R package
itself is in a sub-folder within the project.}

\item{template}{The path to a Quarto file that can be used as the template
for all of the resulting reference files. If left NULL, \code{pkgsite} will use
its default template.}
}
\value{
A character vector with the resulting contents creating document
that links to the converted Quarto documents.
}
\description{
Create a Quarto file that lists the available reference pages
}
\seealso{
Other Quarto file creation: 
\code{\link{rd_to_qmd}()},
\code{\link{write_reference}()},
\code{\link{write_reference_index}()},
\code{\link{write_reference_pages}()}
}
\concept{Quarto file creation}
