% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_to_qmd.R
\name{index_to_qmd}
\alias{index_to_qmd}
\title{Create a Quarto file that lists the available reference pages}
\usage{
index_to_qmd(pkg = ".", template = NULL)
}
\arguments{
\item{pkg}{The path to the package}

\item{template}{The path to a Quarto file that can be used as the template
for all of the resulting reference files. If left NULL, \code{pkgsite} will use
its default template.}
}
\description{
Create a Quarto file that lists the available reference pages
}
