% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-to-list.R
\name{rd_to_list}
\alias{rd_to_list}
\title{Converts a given 'Rd' file into a list object}
\usage{
rd_to_list(file_in, pkg = ".")
}
\arguments{
\item{file_in}{The name of the source Rd file}

\item{pkg}{The path to the package or a \code{pkgdown} package object}
}
\description{
This function is meant to be used as a intermediate object
that could be used as an easy way to convert the information inside the 'Rd'
into other formats or outputs
}
