% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-to-qmd.R
\name{rd_to_qmd}
\alias{rd_to_qmd}
\title{Converts 'Rd' to Quarto files}
\usage{
rd_to_qmd(
  path,
  project = ".",
  pkg = NULL,
  examples = TRUE,
  not_run_examples = FALSE,
  template = NULL
)
}
\arguments{
\item{path}{The path to the source Rd file}

\item{project}{The path to the root folder of the project.}

\item{pkg}{The path inside the project folder. Use only if the R package
itself is in a sub-folder within the project.}

\item{examples}{Flag that sets the examples code chunk to be evaluated when
the Quarto document is rendered}

\item{not_run_examples}{Flag that sets the "do not run" examples code chunk
to be evaluated when the Quarto document is rendered}

\item{template}{The path to a Quarto file that can be used as the template
for all of the resulting reference files. If left NULL, \code{pkgsite} will use
its default template.}
}
\value{
A character vector with the resulting contents of converting the
Rd file format into a Quarto file format.
}
\description{
Converts 'Rd' to Quarto files
}
\examples{
library(pkgsite)
example_pkg <- system.file("example", package = "pkgsite")
rd_to_qmd(file.path(example_pkg, "man", "rd_to_qmd.Rd"), project = example_pkg)
}
\seealso{
Other Conversion functions:
\code{\link[=index_to_qmd]{index_to_qmd()}},
\code{\link[=rd_to_list]{rd_to_list()}}
}
\concept{Conversion functions}
