% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-to-qmd.R
\name{rd_to_qmd}
\alias{rd_to_qmd}
\title{Converts 'Rd' to Quarto files}
\usage{
rd_to_qmd(
  file_in,
  pkg = ".",
  examples = TRUE,
  not_run_examples = FALSE,
  template = NULL
)
}
\arguments{
\item{file_in}{The name of the source Rd file}

\item{pkg}{The path to the package or a \code{pkgdown} package object}

\item{examples}{Flag that sets the examples code chuck to be evaluated when
the Quarto document is rendered}

\item{not_run_examples}{Flag that sets the "do not run" examples code chuck
to be evaluated when the Quarto document is rendered}

\item{template}{The path to a Quarto file that can be used as the template
for all of the resulting reference files. If left NULL, \code{pkgsite} will use
its default template.}
}
\description{
Converts 'Rd' to Quarto files
}
