% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-reference.R
\name{write_reference}
\alias{write_reference}
\title{Writes the reference pages and index as Quarto files}
\usage{
write_reference(
  pkg = NULL,
  folder = NULL,
  examples = TRUE,
  not_run_examples = NULL,
  template = NULL,
  index_file = NULL,
  index_template = NULL
)
}
\arguments{
\item{pkg}{The path to the package or a \code{pkgdown} package object}

\item{folder}{The target folder to save the new Quarto files to}

\item{examples}{Flag that sets the examples code chuck to be evaluated when
the Quarto document is rendered}

\item{not_run_examples}{Flag that sets the "do not run" examples code chuck
to be evaluated when the Quarto document is rendered}

\item{template}{The path to a Quarto file that can be used as the template
for all of the resulting reference files. If left NULL, \code{pkgsite} will use
its default template.}

\item{index_file}{The name assigned to the resulting index Quarto file}

\item{index_template}{The path to a Quarto file that can be used as the
template for the index Quarto file}
}
\description{
Writes the reference pages and index as Quarto files
}
