% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-reference.R
\name{write_reference_index}
\alias{write_reference_index}
\title{Writes the index of the reference pages into a Quarto file}
\usage{
write_reference_index(
  pkg = ".",
  folder = "reference",
  index_file = "index.qmd",
  index_template = NULL
)
}
\arguments{
\item{pkg}{The path to the package}

\item{folder}{The target folder to save the new Quarto files to}

\item{index_file}{The name assigned to the resulting index Quarto file}

\item{index_template}{The path to a Quarto file that can be used as the
template for the index Quarto file}
}
\description{
Writes the index of the reference pages into a Quarto file
}
