% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-reference.R
\name{write_reference_index}
\alias{write_reference_index}
\title{Writes the index of the reference pages into a Quarto file}
\usage{
write_reference_index(
  project = ".",
  pkg = NULL,
  folder = "reference",
  index_file = "index.qmd",
  index_template = NULL
)
}
\arguments{
\item{project}{The path to the root folder of the project.}

\item{pkg}{The path inside the project folder. Use only if the R package
itself is in a sub-folder within the project.}

\item{folder}{The target folder to save the new Quarto files to}

\item{index_file}{The name assigned to the resulting index Quarto file}

\item{index_template}{The path to a Quarto file that can be used as the
template for the index Quarto file}
}
\value{
A Quarto file that links to the individual Quarto documentation
files
}
\description{
Writes the index of the reference pages into a Quarto file
}
\seealso{
Other Quarto file creation: 
\code{\link{write_reference}()},
\code{\link{write_reference_pages}()}
}
\concept{Quarto file creation}
