% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-reference.R
\name{write_reference_pages}
\alias{write_reference_pages}
\title{Converts the 'Rd' file into Quarto, and writes the file to a specified folder}
\usage{
write_reference_pages(
  project = ".",
  pkg = NULL,
  folder = "reference",
  examples = TRUE,
  not_run_examples = FALSE,
  template = NULL
)
}
\arguments{
\item{project}{The path to the root folder of the project.}

\item{pkg}{The path inside the project folder. Use only if the R package
itself is in a sub-folder within the project.}

\item{folder}{The target folder to save the new Quarto files to}

\item{examples}{Flag that sets the examples code chuck to be evaluated when
the Quarto document is rendered}

\item{not_run_examples}{Flag that sets the "do not run" examples code chuck
to be evaluated when the Quarto document is rendered}

\item{template}{The path to a Quarto file that can be used as the template
for all of the resulting reference files. If left NULL, \code{pkgsite} will use
its default template.}
}
\value{
A set of Quarto files written to the specified path based on the
number of Rd files in the package.
}
\description{
Converts the 'Rd' file into Quarto, and writes the file to a specified folder
}
\seealso{
Other Quarto file creation: 
\code{\link{index_to_qmd}()},
\code{\link{rd_to_qmd}()},
\code{\link{write_reference}()},
\code{\link{write_reference_index}()}
}
\concept{Quarto file creation}
